requirements.txt
beautifulsoup4==4.11.1
google==3.0.0
grpcio==1.44.0
grpcio-tools==1.44.0
pkg_resources==0.0.0
protobuf==3.20.1
python-dotenv==0.20.0
six==1.16.0
soupsieve==2.3.2.post1
